hconfig.scm: do not distribute as a source file.
hconfig.scm should be built as part of the build process. It is derived
from hconfig.scm.hall.

As per
https://www.gnu.org/software/automake/manual/html_node/Errors-with-distclean.html,
the file is intended to always be rebuilt by users, that is by the
person installing the project managed by hall for eventual use, during
the installation process. So it should not be part of the distribution.