Micro fix: `-f` flag for `guix shell` in README.org.
In README.org, `guix shell` should be using the `-f` or `--file` instead of
`-l`, which is a legacy from `guix environment`.

Also, when used to create a development environment, `guix shell` should be
called with the `-D` or `--development` option. The command's first occurrence
in the README.org file has been amended to reflect this. A link to the Guix
official documentation has also been added.